Fix compile warnings
Fix two warnings:

e:\xeniface\src\xenagent\service.cpp(684): warning C26439: This kind offunction may not throw.
Declare it 'noexcept' (f.6). [E:\xeniface\vs2017\xenagent\xenagent.vcxproj]

e:\xeniface\src\xenagent\xenifacedevice.cpp(57): warning C26451: Arithmetic overflow: Using operator
'+' on a 4 byte value and then casting the result to a 8 byte value. Cast the value to the wider type before
calling operator '+' to avoid overflow (io.2). [E:\xeniface\vs2017\xenagent\xenagent.vcxproj]

+++ src/xenagent/xenifacedevice.cpp
@@ -54,7 +54,7 @@ bool CXenIfaceDevice::StoreRead(const std::string& path, std::string& value)
           NULL, 0,
           &bytes);
 
-    buffer = new char[bytes + 1];
+    buffer = new char[(size_t)bytes + 1];
     if (buffer == NULL)
         return false;
